refactor: Carve out type definitions
And introduce `TransientOutputPath` for atomic-renamed output
